% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVMR-cML-SuSiE-package.R
\docType{package}
\name{MVMR-cML-SuSiE-package}
\alias{MVMR-cML-SuSiE}
\alias{MVMR-cML-SuSiE-package}
\title{MVMR-cML-SuSiE: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
