% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mvmr.cml.susie.step1}
\alias{mvmr.cml.susie.step1}
\title{Step 1 of MVMR-cML-SuSiE}
\usage{
mvmr.cml.susie.step1(
  exposure.ids = NULL,
  outcome.id = NULL,
  sample.sizes,
  beta.exposure.ls = NULL,
  se.exposure.ls = NULL,
  beta.outcome.ls = NULL,
  se.outcome.ls = NULL,
  use.openGWAS = TRUE
)
}
\arguments{
\item{exposure.ids}{A length \eqn{L} character vector of openGWAS IDs corresponding to the exposures,
where \eqn{L} is the number of intended exposures}

\item{outcome.id}{A character corresponding to the outcome openGWAS ID}

\item{sample.sizes}{A numeric vector of sample sizes corresponding to each exposure and outcome
If \code{use.openGWAS = TRUE}, the length should be \eqn{L} (exposures only).
Otherwise, the length should be \eqn{L + 1}, with the last element for the outcome sample size}

\item{beta.exposure.ls}{A list (length \eqn{L}) of numeric vectors with effect sizes (betas) for each SNP on exposure}

\item{se.exposure.ls}{A list (length \eqn{L}) of numeric vectors with standard errors for each SNP on exposure}

\item{beta.outcome.ls}{A list (length \eqn{L}) of numeric vectors with effect sizes (betas) for each SNP on outcome}

\item{se.outcome.ls}{A list (length \eqn{L}) of numeric vectors with standard errors for each SNP on outcome}

\item{use.openGWAS}{A logical indicating whether to use \code{OpenGWAS} to extract GWAS data based on
\code{exposure.ids} and \code{outcome.id}. If \code{FALSE}, requires user-provided summary statistics
(lists for \code{beta.exposure.ls}, \code{se.exposure.ls}, \code{beta.outcome.ls}, and \code{se.outcome.ls})
Default is \code{TRUE}}
}
\value{
A numeric vector (length \eqn{L}) of p-values, each corresponding to the UVMR p-value for one exposure
}
\description{
This function performs the first step of the MVMR-cML-SuSiE approach to obtain each exposure's univariable
Mendelian randomization (UVMR) p-value, used for initial filtering of exposures. You can choose to pull
GWAS data using \code{OpenGWAS} (by setting \code{use.openGWAS = TRUE}) or provide your own summary
statistics (by setting \code{use.openGWAS = FALSE})
}
