% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mvmr.cml.susie.step2}
\alias{mvmr.cml.susie.step2}
\title{Step 2 of MVMR-cML-SuSiE}
\usage{
mvmr.cml.susie.step2(
  exposure.ids.subset = NULL,
  outcome.id = NULL,
  sample.sizes.subset,
  beta.exposure.mat = NULL,
  se.exposure.mat = NULL,
  beta.outcome.vec = NULL,
  se.outcome.vec = NULL,
  pval.exposure.mat = NULL,
  use.openGWAS = TRUE,
  cutoff = 5e-08
)
}
\arguments{
\item{exposure.ids.subset}{A length \eqn{L^{*}} character vector of exposure IDs (a subset from step 1)}

\item{outcome.id}{A character specifying the outcome ID}

\item{sample.sizes.subset}{A length \eqn{L^{*}} numeric vector of sample sizes corresponding to
\code{exposure.ids.subset} (and optionally the outcome as the last element if user-provided data)}

\item{beta.exposure.mat}{A size \eqn{m^{*} \times L^{*}} matrix of exposure effect sizes, where
\eqn{m^{*}} is the number of SNPs}

\item{se.exposure.mat}{A size \eqn{m^{*} \times L^{*}} matrix of exposure standard errors}

\item{beta.outcome.vec}{A length \eqn{m^{*}} numeric vector of outcome effect sizes for the same SNPs}

\item{se.outcome.vec}{A length \eqn{m^{*}} numeric vector of standard errors for \code{beta.outcome.vec}}

\item{pval.exposure.mat}{A \eqn{m^{*} \times L^{*}} matrix of exposure p-values}

\item{use.openGWAS}{A logical indicating whether to extract data via \code{OpenGWAS}. If \code{FALSE},
you must supply the \code{*.mat} and \code{*.vec} arguments. Default is \code{TRUE}}

\item{cutoff}{A numeric threshold for instrument selection. Default is \eqn{5 \times 10^{-8}}}
}
\value{
A list containing:
\item{mvdat}{A list containing the relevant data (exposure/outcome betas, SEs, etc)}
\item{invalid.idx}{A vector of indices of invalid instruments identified based off the UVMR procedure}
\item{theta.vec}{A numeric vector of initial values for each exposure}
}
\description{
This function performs the second step of the MVMR-cML-SuSiE approach. After identifying
exposures of interest in Step 1 (e.g., by filtering UVMR p-values), step 2 extracts the
relevant instruments for those exposures, identifies invalid instruments (if any),
and obtains an initial estimate from univariable Mendelian randomization (UVMR) for each exposure
}
