% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mvmr.cml.susie.step3}
\alias{mvmr.cml.susie.step3}
\title{Step 3 of MVMR-cML-SuSiE (Iterative SuSiE)}
\usage{
mvmr.cml.susie.step3(
  mvdat,
  invalid.idx,
  theta.vec,
  rho.mat,
  S = 10,
  max.iter = 200,
  tol = 1e-10
)
}
\arguments{
\item{mvdat}{A list returned by \code{mvmr.cml.susie.step2}, containing exposure and outcome summary statistics}

\item{invalid.idx}{An integer vector of invalid IV indices as identified in step 2}

\item{theta.vec}{A numeric vector of initial exposure effect estimates from step 2}

\item{rho.mat}{A genetic correlation matrix of size \eqn{(L^{*} + 1) \times (L^{*} + 1)}, where \eqn{L^{*}} is
the number of exposures. The last row/column corresponds to the outcome}

\item{S}{The number of single effects (signal clusters) in SuSiE. Default is 10.}

\item{max.iter}{The maximum number of iterations for SuSiE. Default is 200.}

\item{tol}{A numeric value indicating the convergence threshold for the iterative procedure. Default is \eqn{1 \times 10^{-10}}.}
}
\value{
A fitted \code{susie} object. The key SuSiE output is the posterior inclusion probability (PIP) matrix \code{alpha}
used for determining signal clusters
}
\description{
This function performs the iterative SuSiE algorithm to identify potential exposure signal clusters
}
